LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/scale-interpolation-expected.txt


FAIL CSS Transitions: property <scale> from [-10 5] to [10 -5] at (-1) should be [-30 15] assert_equals: expected "- 30 15 " but got "10 - 5 "
FAIL CSS Transitions: property <scale> from [-10 5] to [10 -5] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "10 - 5 "
FAIL CSS Transitions: property <scale> from [-10 5] to [10 -5] at (0.25) should be [-5 2.5] assert_equals: expected "- 5 2.5 " but got "10 - 5 "
FAIL CSS Transitions: property <scale> from [-10 5] to [10 -5] at (0.75) should be [5 -2.5] assert_equals: expected "5 - 2.5 " but got "10 - 5 "
PASS CSS Transitions: property <scale> from [-10 5] to [10 -5] at (1) should be [10 -5]
FAIL CSS Transitions: property <scale> from [-10 5] to [10 -5] at (2) should be [30 -15] assert_equals: expected "30 - 15 " but got "10 - 5 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (-1) should be [-30 15] assert_equals: expected "- 30 15 " but got "10 - 5 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "10 - 5 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (0.25) should be [-5 2.5] assert_equals: expected "- 5 2.5 " but got "10 - 5 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (0.75) should be [5 -2.5] assert_equals: expected "5 - 2.5 " but got "10 - 5 "
PASS CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (1) should be [10 -5]
FAIL CSS Transitions with transition: all: property <scale> from [-10 5] to [10 -5] at (2) should be [30 -15] assert_equals: expected "30 - 15 " but got "10 - 5 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (-1) should be [-30 15] assert_equals: expected "- 30 15 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (0.25) should be [-5 2.5] assert_equals: expected "- 5 2.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (0.75) should be [5 -2.5] assert_equals: expected "5 - 2.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (1) should be [10 -5] assert_equals: expected "10 - 5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5] to [10 -5] at (2) should be [30 -15] assert_equals: expected "30 - 15 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (-1) should be [-30 15] assert_equals: expected "- 30 15 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (0.25) should be [-5 2.5] assert_equals: expected "- 5 2.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (0.75) should be [5 -2.5] assert_equals: expected "5 - 2.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (1) should be [10 -5] assert_equals: expected "10 - 5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5] to [10 -5] at (2) should be [30 -15] assert_equals: expected "30 - 15 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (-1) should be [-6 -50 -400] assert_equals: expected "- 6 - 50 - 400 " but got "10 110 1200 "
FAIL CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (0) should be [2 30 400] assert_equals: expected "2 30 400 " but got "10 110 1200 "
FAIL CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (0.125) should be [3 40 500] assert_equals: expected "3 40 500 " but got "10 110 1200 "
FAIL CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (0.875) should be [9 100 1100] assert_equals: expected "9 100 1100 " but got "10 110 1200 "
PASS CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (1) should be [10 110 1200]
FAIL CSS Transitions: property <scale> from [2 30 400] to [10 110 1200] at (2) should be [18 190 2000] assert_equals: expected "18 190 2000 " but got "10 110 1200 "
FAIL CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (-1) should be [-6 -50 -400] assert_equals: expected "- 6 - 50 - 400 " but got "10 110 1200 "
FAIL CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (0) should be [2 30 400] assert_equals: expected "2 30 400 " but got "10 110 1200 "
FAIL CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (0.125) should be [3 40 500] assert_equals: expected "3 40 500 " but got "10 110 1200 "
FAIL CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (0.875) should be [9 100 1100] assert_equals: expected "9 100 1100 " but got "10 110 1200 "
PASS CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (1) should be [10 110 1200]
FAIL CSS Transitions with transition: all: property <scale> from [2 30 400] to [10 110 1200] at (2) should be [18 190 2000] assert_equals: expected "18 190 2000 " but got "10 110 1200 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (-1) should be [-6 -50 -400] assert_equals: expected "- 6 - 50 - 400 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (0) should be [2 30 400] assert_equals: expected "2 30 400 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (0.125) should be [3 40 500] assert_equals: expected "3 40 500 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (0.875) should be [9 100 1100] assert_equals: expected "9 100 1100 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (1) should be [10 110 1200] assert_equals: expected "10 110 1200 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 30 400] to [10 110 1200] at (2) should be [18 190 2000] assert_equals: expected "18 190 2000 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (-1) should be [-6 -50 -400] assert_equals: expected "- 6 - 50 - 400 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (0) should be [2 30 400] assert_equals: expected "2 30 400 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (0.125) should be [3 40 500] assert_equals: expected "3 40 500 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (0.875) should be [9 100 1100] assert_equals: expected "9 100 1100 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (1) should be [10 110 1200] assert_equals: expected "10 110 1200 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 30 400] to [10 110 1200] at (2) should be [18 190 2000] assert_equals: expected "18 190 2000 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [26 17 9] to [2 1] at (-1) should be [50 33 17] assert_equals: expected "50 33 17 " but got "2 1 "
FAIL CSS Transitions: property <scale> from [26 17 9] to [2 1] at (0) should be [26 17 9] assert_equals: expected "26 17 9 " but got "2 1 "
FAIL CSS Transitions: property <scale> from [26 17 9] to [2 1] at (0.125) should be [23 15 8] assert_equals: expected "23 15 8 " but got "2 1 "
FAIL CSS Transitions: property <scale> from [26 17 9] to [2 1] at (0.875) should be [5 3 2] assert_equals: expected "5 3 2 " but got "2 1 "
PASS CSS Transitions: property <scale> from [26 17 9] to [2 1] at (1) should be [2 1]
FAIL CSS Transitions: property <scale> from [26 17 9] to [2 1] at (2) should be [-22 -15 -7] assert_equals: expected "- 22 - 15 - 7 " but got "2 1 "
FAIL CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (-1) should be [50 33 17] assert_equals: expected "50 33 17 " but got "2 1 "
FAIL CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (0) should be [26 17 9] assert_equals: expected "26 17 9 " but got "2 1 "
FAIL CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (0.125) should be [23 15 8] assert_equals: expected "23 15 8 " but got "2 1 "
FAIL CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (0.875) should be [5 3 2] assert_equals: expected "5 3 2 " but got "2 1 "
PASS CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (1) should be [2 1]
FAIL CSS Transitions with transition: all: property <scale> from [26 17 9] to [2 1] at (2) should be [-22 -15 -7] assert_equals: expected "- 22 - 15 - 7 " but got "2 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (-1) should be [50 33 17] assert_equals: expected "50 33 17 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (0) should be [26 17 9] assert_equals: expected "26 17 9 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (0.125) should be [23 15 8] assert_equals: expected "23 15 8 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (0.875) should be [5 3 2] assert_equals: expected "5 3 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (1) should be [2 1] assert_equals: expected "2 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [26 17 9] to [2 1] at (2) should be [-22 -15 -7] assert_equals: expected "- 22 - 15 - 7 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (-1) should be [50 33 17] assert_equals: expected "50 33 17 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (0) should be [26 17 9] assert_equals: expected "26 17 9 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (0.125) should be [23 15 8] assert_equals: expected "23 15 8 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (0.875) should be [5 3 2] assert_equals: expected "5 3 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (1) should be [2 1] assert_equals: expected "2 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [26 17 9] to [2 1] at (2) should be [-22 -15 -7] assert_equals: expected "- 22 - 15 - 7 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [1] to [10 -5 0] at (-1) should be [-8 7 2] assert_equals: expected "- 8 7 2 " but got "10 - 5 0 "
FAIL CSS Transitions: property <scale> from [1] to [10 -5 0] at (0) should be [1] assert_equals: expected "1 1 " but got "10 - 5 0 "
FAIL CSS Transitions: property <scale> from [1] to [10 -5 0] at (0.25) should be [3.25 -0.5 0.75] assert_equals: expected "3.25 - 0.5 0.75 " but got "10 - 5 0 "
FAIL CSS Transitions: property <scale> from [1] to [10 -5 0] at (0.75) should be [7.75 -3.5 0.25] assert_equals: expected "7.75 - 3.5 0.25 " but got "10 - 5 0 "
PASS CSS Transitions: property <scale> from [1] to [10 -5 0] at (1) should be [10 -5 0]
FAIL CSS Transitions: property <scale> from [1] to [10 -5 0] at (2) should be [19 -11 -1] assert_equals: expected "19 - 11 - 1 " but got "10 - 5 0 "
FAIL CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (-1) should be [-8 7 2] assert_equals: expected "- 8 7 2 " but got "10 - 5 0 "
FAIL CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (0) should be [1] assert_equals: expected "1 1 " but got "10 - 5 0 "
FAIL CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (0.25) should be [3.25 -0.5 0.75] assert_equals: expected "3.25 - 0.5 0.75 " but got "10 - 5 0 "
FAIL CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (0.75) should be [7.75 -3.5 0.25] assert_equals: expected "7.75 - 3.5 0.25 " but got "10 - 5 0 "
PASS CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (1) should be [10 -5 0]
FAIL CSS Transitions with transition: all: property <scale> from [1] to [10 -5 0] at (2) should be [19 -11 -1] assert_equals: expected "19 - 11 - 1 " but got "10 - 5 0 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (-1) should be [-8 7 2] assert_equals: expected "- 8 7 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (0.25) should be [3.25 -0.5 0.75] assert_equals: expected "3.25 - 0.5 0.75 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (0.75) should be [7.75 -3.5 0.25] assert_equals: expected "7.75 - 3.5 0.25 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (1) should be [10 -5 0] assert_equals: expected "10 - 5 0 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [1] to [10 -5 0] at (2) should be [19 -11 -1] assert_equals: expected "19 - 11 - 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (-1) should be [-8 7 2] assert_equals: expected "- 8 7 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (0.25) should be [3.25 -0.5 0.75] assert_equals: expected "3.25 - 0.5 0.75 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (0.75) should be [7.75 -3.5 0.25] assert_equals: expected "7.75 - 3.5 0.25 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (1) should be [10 -5 0] assert_equals: expected "10 - 5 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [1] to [10 -5 0] at (2) should be [19 -11 -1] assert_equals: expected "19 - 11 - 1 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [-10 5 1] to [1] at (-1) should be [-21 9] assert_equals: expected "- 21 9 " but got "1 1 "
FAIL CSS Transitions: property <scale> from [-10 5 1] to [1] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1 1 "
FAIL CSS Transitions: property <scale> from [-10 5 1] to [1] at (0.25) should be [-7.25 4] assert_equals: expected "- 7.25 4 " but got "1 1 "
FAIL CSS Transitions: property <scale> from [-10 5 1] to [1] at (0.75) should be [-1.75 2] assert_equals: expected "- 1.75 2 " but got "1 1 "
PASS CSS Transitions: property <scale> from [-10 5 1] to [1] at (1) should be [1]
FAIL CSS Transitions: property <scale> from [-10 5 1] to [1] at (2) should be [12 -3] assert_equals: expected "12 - 3 " but got "1 1 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (-1) should be [-21 9] assert_equals: expected "- 21 9 " but got "1 1 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1 1 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (0.25) should be [-7.25 4] assert_equals: expected "- 7.25 4 " but got "1 1 "
FAIL CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (0.75) should be [-1.75 2] assert_equals: expected "- 1.75 2 " but got "1 1 "
PASS CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (1) should be [1]
FAIL CSS Transitions with transition: all: property <scale> from [-10 5 1] to [1] at (2) should be [12 -3] assert_equals: expected "12 - 3 " but got "1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (-1) should be [-21 9] assert_equals: expected "- 21 9 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (0.25) should be [-7.25 4] assert_equals: expected "- 7.25 4 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (0.75) should be [-1.75 2] assert_equals: expected "- 1.75 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [-10 5 1] to [1] at (2) should be [12 -3] assert_equals: expected "12 - 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (-1) should be [-21 9] assert_equals: expected "- 21 9 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (0) should be [-10 5] assert_equals: expected "- 10 5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (0.25) should be [-7.25 4] assert_equals: expected "- 7.25 4 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (0.75) should be [-1.75 2] assert_equals: expected "- 1.75 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [-10 5 1] to [1] at (2) should be [12 -3] assert_equals: expected "12 - 3 " but got "1.1 1 "
PASS CSS Transitions: property <scale> from [none] to [none] at (-1) should be [none]
PASS CSS Transitions: property <scale> from [none] to [none] at (0) should be [none]
PASS CSS Transitions: property <scale> from [none] to [none] at (0.125) should be [none]
PASS CSS Transitions: property <scale> from [none] to [none] at (0.875) should be [none]
PASS CSS Transitions: property <scale> from [none] to [none] at (1) should be [none]
PASS CSS Transitions: property <scale> from [none] to [none] at (2) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (-1) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (0) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (0.125) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (0.875) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (1) should be [none]
PASS CSS Transitions with transition: all: property <scale> from [none] to [none] at (2) should be [none]
FAIL CSS Animations: property <scale> from [none] to [none] at (-1) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [none] at (0) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [none] at (0.125) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [none] at (0.875) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [none] at (1) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [none] at (2) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (-1) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (0) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (0.125) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (0.875) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (1) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [none] at (2) should be [none] assert_equals: expected "none " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [none] to [4 3 2] at (-1) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "4 3 2 "
FAIL CSS Transitions: property <scale> from [none] to [4 3 2] at (0) should be [1] assert_equals: expected "1 1 " but got "4 3 2 "
FAIL CSS Transitions: property <scale> from [none] to [4 3 2] at (0.125) should be [1.375 1.25 1.125] assert_equals: expected "1.38 1.25 1.13 " but got "4 3 2 "
FAIL CSS Transitions: property <scale> from [none] to [4 3 2] at (0.875) should be [3.625 2.75 1.875] assert_equals: expected "3.63 2.75 1.88 " but got "4 3 2 "
PASS CSS Transitions: property <scale> from [none] to [4 3 2] at (1) should be [4 3 2]
FAIL CSS Transitions: property <scale> from [none] to [4 3 2] at (2) should be [7 5 3] assert_equals: expected "7 5 3 " but got "4 3 2 "
FAIL CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (-1) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "4 3 2 "
FAIL CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (0) should be [1] assert_equals: expected "1 1 " but got "4 3 2 "
FAIL CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (0.125) should be [1.375 1.25 1.125] assert_equals: expected "1.38 1.25 1.13 " but got "4 3 2 "
FAIL CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (0.875) should be [3.625 2.75 1.875] assert_equals: expected "3.63 2.75 1.88 " but got "4 3 2 "
PASS CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (1) should be [4 3 2]
FAIL CSS Transitions with transition: all: property <scale> from [none] to [4 3 2] at (2) should be [7 5 3] assert_equals: expected "7 5 3 " but got "4 3 2 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (-1) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (0.125) should be [1.375 1.25 1.125] assert_equals: expected "1.38 1.25 1.13 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (0.875) should be [3.625 2.75 1.875] assert_equals: expected "3.63 2.75 1.88 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (1) should be [4 3 2] assert_equals: expected "4 3 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [none] to [4 3 2] at (2) should be [7 5 3] assert_equals: expected "7 5 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (-1) should be [-2 -1 0] assert_equals: expected "- 2 - 1 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (0.125) should be [1.375 1.25 1.125] assert_equals: expected "1.38 1.25 1.13 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (0.875) should be [3.625 2.75 1.875] assert_equals: expected "3.63 2.75 1.88 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (1) should be [4 3 2] assert_equals: expected "4 3 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [none] to [4 3 2] at (2) should be [7 5 3] assert_equals: expected "7 5 3 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from neutral to [1.5 1] at (-1) should be [0.7 1] assert_equals: expected "0.7 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from neutral to [1.5 1] at (0) should be [1.1 1] assert_equals: expected "1.1 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from neutral to [1.5 1] at (0.25) should be [1.2 1] assert_equals: expected "1.2 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from neutral to [1.5 1] at (0.75) should be [1.4 1] assert_equals: expected "1.4 1 " but got "1.5 1 "
PASS CSS Transitions: property <scale> from neutral to [1.5 1] at (1) should be [1.5 1]
FAIL CSS Transitions: property <scale> from neutral to [1.5 1] at (2) should be [1.9 1] assert_equals: expected "1.9 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (-1) should be [0.7 1] assert_equals: expected "0.7 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (0) should be [1.1 1] assert_equals: expected "1.1 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (0.25) should be [1.2 1] assert_equals: expected "1.2 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (0.75) should be [1.4 1] assert_equals: expected "1.4 1 " but got "1.5 1 "
PASS CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (1) should be [1.5 1]
FAIL CSS Transitions with transition: all: property <scale> from neutral to [1.5 1] at (2) should be [1.9 1] assert_equals: expected "1.9 1 " but got "1.5 1 "
FAIL CSS Animations: property <scale> from neutral to [1.5 1] at (-1) should be [0.7 1] assert_equals: expected "0.7 1 " but got "1.1 1 "
PASS CSS Animations: property <scale> from neutral to [1.5 1] at (0) should be [1.1 1]
FAIL CSS Animations: property <scale> from neutral to [1.5 1] at (0.25) should be [1.2 1] assert_equals: expected "1.2 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from neutral to [1.5 1] at (0.75) should be [1.4 1] assert_equals: expected "1.4 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from neutral to [1.5 1] at (1) should be [1.5 1] assert_equals: expected "1.5 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from neutral to [1.5 1] at (2) should be [1.9 1] assert_equals: expected "1.9 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from neutral to [1.5 1] at (-1) should be [0.7 1] assert_equals: expected "0.7 1 " but got "1.1 1 "
PASS Web Animations: property <scale> from neutral to [1.5 1] at (0) should be [1.1 1]
FAIL Web Animations: property <scale> from neutral to [1.5 1] at (0.25) should be [1.2 1] assert_equals: expected "1.2 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from neutral to [1.5 1] at (0.75) should be [1.4 1] assert_equals: expected "1.4 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from neutral to [1.5 1] at (1) should be [1.5 1] assert_equals: expected "1.5 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from neutral to [1.5 1] at (2) should be [1.9 1] assert_equals: expected "1.9 1 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (-1) should be [0 1.5] assert_equals: expected "0 1.5 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (0.25) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (0.75) should be [1.75 0.625] assert_equals: expected "1.75 0.63 " but got "2 0.5 "
PASS CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (1) should be [2 0.5]
FAIL CSS Transitions: property <scale> from [initial] to [2 0.5 1] at (2) should be [3 0] assert_equals: expected "3 0 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (-1) should be [0 1.5] assert_equals: expected "0 1.5 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (0.25) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (0.75) should be [1.75 0.625] assert_equals: expected "1.75 0.63 " but got "2 0.5 "
PASS CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (1) should be [2 0.5]
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [2 0.5 1] at (2) should be [3 0] assert_equals: expected "3 0 " but got "2 0.5 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (-1) should be [0 1.5] assert_equals: expected "0 1.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (0.25) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (0.75) should be [1.75 0.625] assert_equals: expected "1.75 0.63 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (1) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [2 0.5 1] at (2) should be [3 0] assert_equals: expected "3 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (-1) should be [0 1.5] assert_equals: expected "0 1.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (0.25) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (0.75) should be [1.75 0.625] assert_equals: expected "1.75 0.63 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (1) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [2 0.5 1] at (2) should be [3 0] assert_equals: expected "3 0 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (-1) should be [3 0] assert_equals: expected "3 0 " but got "none "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "none "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (0.25) should be [1.75 0.6251] assert_equals: expected "1.75 0.63 " but got "none "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (0.75) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "none "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "none "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [initial] at (2) should be [0 1.5] assert_equals: expected "0 1.5 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (-1) should be [3 0] assert_equals: expected "3 0 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (0.25) should be [1.75 0.6251] assert_equals: expected "1.75 0.63 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (0.75) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [initial] at (2) should be [0 1.5] assert_equals: expected "0 1.5 " but got "none "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (-1) should be [3 0] assert_equals: expected "3 0 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (0.25) should be [1.75 0.6251] assert_equals: expected "1.75 0.63 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (0.75) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [initial] at (2) should be [0 1.5] assert_equals: expected "0 1.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (-1) should be [3 0] assert_equals: expected "3 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (0.25) should be [1.75 0.6251] assert_equals: expected "1.75 0.63 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (0.75) should be [1.25 0.875] assert_equals: expected "1.25 0.88 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [initial] at (2) should be [0 1.5] assert_equals: expected "0 1.5 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [unset] to [1.5 1] at (-1) should be [0.5 1] assert_equals: expected "0.5 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from [unset] to [1.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from [unset] to [1.5 1] at (0.25) should be [1.125 1] assert_equals: expected "1.13 1 " but got "1.5 1 "
FAIL CSS Transitions: property <scale> from [unset] to [1.5 1] at (0.75) should be [1.375 1] assert_equals: expected "1.38 1 " but got "1.5 1 "
PASS CSS Transitions: property <scale> from [unset] to [1.5 1] at (1) should be [1.5 1]
FAIL CSS Transitions: property <scale> from [unset] to [1.5 1] at (2) should be [2 1] assert_equals: expected "2 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (-1) should be [0.5 1] assert_equals: expected "0.5 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (0.25) should be [1.125 1] assert_equals: expected "1.13 1 " but got "1.5 1 "
FAIL CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (0.75) should be [1.375 1] assert_equals: expected "1.38 1 " but got "1.5 1 "
PASS CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (1) should be [1.5 1]
FAIL CSS Transitions with transition: all: property <scale> from [unset] to [1.5 1] at (2) should be [2 1] assert_equals: expected "2 1 " but got "1.5 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (-1) should be [0.5 1] assert_equals: expected "0.5 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (0.25) should be [1.125 1] assert_equals: expected "1.13 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (0.75) should be [1.375 1] assert_equals: expected "1.38 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (1) should be [1.5 1] assert_equals: expected "1.5 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [unset] to [1.5 1] at (2) should be [2 1] assert_equals: expected "2 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (-1) should be [0.5 1] assert_equals: expected "0.5 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (0.25) should be [1.125 1] assert_equals: expected "1.13 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (0.75) should be [1.375 1] assert_equals: expected "1.38 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (1) should be [1.5 1] assert_equals: expected "1.5 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [unset] to [1.5 1] at (2) should be [2 1] assert_equals: expected "2 1 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (-1) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (0.25) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "2 0.5 "
FAIL CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (0.75) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "2 0.5 "
PASS CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (1) should be [2 0.5]
FAIL CSS Transitions: property <scale> from [inherit] to [2 0.5 1] at (2) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (-1) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (0.25) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "2 0.5 "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (0.75) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "2 0.5 "
PASS CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (1) should be [2 0.5]
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [2 0.5 1] at (2) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "2 0.5 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (-1) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (0.25) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (0.75) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (1) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [2 0.5 1] at (2) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (-1) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (0.25) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (0.75) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (1) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [2 0.5 1] at (2) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (-1) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (0.25) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (0.75) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "0.5 1 2 "
PASS CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (1) should be [0.5 1 2]
FAIL CSS Transitions: property <scale> from [2 0.5 1] to [inherit] at (2) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (-1) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (0.25) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (0.75) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "0.5 1 2 "
PASS CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (1) should be [0.5 1 2]
FAIL CSS Transitions with transition: all: property <scale> from [2 0.5 1] to [inherit] at (2) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "0.5 1 2 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (-1) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (0.25) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (0.75) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (1) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [2 0.5 1] to [inherit] at (2) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (-1) should be [3.5 0 0] assert_equals: expected "3.5 0 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (0) should be [2 0.5] assert_equals: expected "2 0.5 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (0.25) should be [1.625 0.625 1.25] assert_equals: expected "1.63 0.63 1.25 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (0.75) should be [0.875 0.875 1.75] assert_equals: expected "0.88 0.88 1.75 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (1) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [2 0.5 1] to [inherit] at (2) should be [-1 1.5 3] assert_equals: expected "- 1 1.5 3 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [initial] to [inherit] at (-1) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [initial] to [inherit] at (0) should be [1] assert_equals: expected "1 1 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [initial] to [inherit] at (0.25) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "0.5 1 2 "
FAIL CSS Transitions: property <scale> from [initial] to [inherit] at (0.75) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "0.5 1 2 "
PASS CSS Transitions: property <scale> from [initial] to [inherit] at (1) should be [0.5 1 2]
FAIL CSS Transitions: property <scale> from [initial] to [inherit] at (2) should be [0 1 3] assert_equals: expected "0 1 3 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (-1) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (0) should be [1] assert_equals: expected "1 1 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (0.25) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "0.5 1 2 "
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (0.75) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "0.5 1 2 "
PASS CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (1) should be [0.5 1 2]
FAIL CSS Transitions with transition: all: property <scale> from [initial] to [inherit] at (2) should be [0 1 3] assert_equals: expected "0 1 3 " but got "0.5 1 2 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (-1) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (0.25) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (0.75) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (1) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [initial] to [inherit] at (2) should be [0 1 3] assert_equals: expected "0 1 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (-1) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (0) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (0.25) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (0.75) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (1) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [initial] to [inherit] at (2) should be [0 1 3] assert_equals: expected "0 1 3 " but got "1.1 1 "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (-1) should be [0 1 3] assert_equals: expected "0 1 3 " but got "none "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "none "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (0.25) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "none "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (0.75) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "none "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "none "
FAIL CSS Transitions: property <scale> from [inherit] to [initial] at (2) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (-1) should be [0 1 3] assert_equals: expected "0 1 3 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (0.25) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (0.75) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "none "
FAIL CSS Transitions with transition: all: property <scale> from [inherit] to [initial] at (2) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "none "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (-1) should be [0 1 3] assert_equals: expected "0 1 3 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (0.25) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (0.75) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL CSS Animations: property <scale> from [inherit] to [initial] at (2) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (-1) should be [0 1 3] assert_equals: expected "0 1 3 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (0) should be [0.5 1 2] assert_equals: expected "0.5 1 2 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (0.25) should be [0.625 1 1.75] assert_equals: expected "0.63 1 1.75 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (0.75) should be [0.875 1 1.25] assert_equals: expected "0.88 1 1.25 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (1) should be [1] assert_equals: expected "1 1 " but got "1.1 1 "
FAIL Web Animations: property <scale> from [inherit] to [initial] at (2) should be [1.5 1 0] assert_equals: expected "1.5 1 0 " but got "1.1 1 "
